refactor: 패키지명 오타 수정 및 interceptor 등록 순서 변경
