+++ src/main/java/com/sprint/deokhugam/global/config/WebMvcConfig.java
@@ -1,7 +1,7 @@
 package com.sprint.deokhugam.global.config;
 
-import com.sprint.deokhugam.global.inteceptor.LoginInterceptor;
-import com.sprint.deokhugam.global.inteceptor.MDCLoggingInterceptor;
+import com.sprint.deokhugam.global.interceptor.LoginInterceptor;
+import com.sprint.deokhugam.global.interceptor.MDCLoggingInterceptor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
@@ -16,11 +16,11 @@ public class WebMvcConfig implements WebMvcConfigurer {
 
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(mdcLoggingInterceptor)
+            .addPathPatterns("/api/**");
+
         registry.addInterceptor(loginInterceptor)
                 .addPathPatterns("/api/**")
                     .excludePathPatterns("/api/users/login", "/api/users"); // 로그인 회원가입 요청만 허용
-
-        registry.addInterceptor(mdcLoggingInterceptor)
-            .addPathPatterns("/api/**");
     }
 }
